Intermediate save...pivoting to generating unit_descriptor from data saved in sorted_units
